mqtt/subscriber.py
# python3.6

import random
import sys
from paho.mqtt import client as mqtt_client
import os
import time

broker = '10.102.20.105'
port = 1884
topic = "python/mqtt"
# generate client ID with pub prefix randomly
client_id = f'python-mqtt-{random.randint(0, 100)}'
# username = 'emqx'
# password = 'public'


def connect_mqtt() -> mqtt_client:
    def on_connect(client, userdata, flags, rc):
        if rc == 0:
            print("Connected to MQTT Broker!")
        else:
            print("Failed to connect, return code %d\n", rc)

    client = mqtt_client.Client(client_id)
    client.on_connect = on_connect
    client.connect(broker, port)
    return client


def subscribe(client: mqtt_client):
    def on_message(client, userdata, msg):
        received = msg.payload.decode()
        
        if received != "STOP":
            os.system('clear')
            sys.stdout.write(received)

        else:
            os.system('clear')
            sys.stdout.write("STOP")
            


    client.subscribe(topic)
    client.on_message = on_message


def run():
    client = connect_mqtt()
    subscribe(client)
    client.loop_forever()


if __name__ == '__main__':
    run()
