mqtt/requirements.txt
paho-mqtt
ipywidgets